requirements.txt
fastapi>=0.100,<1.0
uvicorn[standard]>=0.23
pydantic>=2,<3
pydantic-settings>=2,<3
sqlalchemy>=2,<3
asyncpg>=0.28
alembic>=1.12
passlib[bcrypt]>=1.7
bcrypt>=4.0,<5
pyjwt>=2.8
python-multipart>=0.0.9
python-slugify>=8.0
aiofiles>=23.0
loguru>=0.7
slowapi>=0.1.8
redis>=5
celery>=5.3
httpx>=0.25
pytest>=7
pytest-asyncio>=0.21
email-validator>=2.0
fastapi-mail>=1.4
psycopg[binary]>=3.1
alibabacloud-dysmsapi20170525==2.0.24
openai>=1.0
